% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cts_kernel_generator.R
\name{cts_kernel_generator}
\alias{cts_kernel_generator}
\title{Circadian Time of Death from Kernal Density of Real Data}
\usage{
cts_kernel_generator(observed_cts, cts_lb = 0, cts_ub = 24, n)
}
\arguments{
\item{observed_cts}{Observed time of death in real data. Lie between cts_lb and cts_ub.}

\item{cts_lb}{Lower bound of time of death.}

\item{cts_ub}{Lower bound of time of death.}

\item{n}{Number of time of death to generate.}
}
\value{
A time of deach vector of length n
}
\description{
Sample circadian time of death from the density of observed time of death in real data.
}
\details{
Generate circadian time of death at various sample sizes from kernel density estimated
from observed time of death in real data.
}
\examples{
data(brain_data)
cts_kernel_generator(observed_cts=BA11_BA47.TOD, cts_lb = 0, cts_ub = 24, n = 100)
}
\author{
Wei Zong, Zhiguang Huo
}
