% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlPFC.TOD.R
\docType{data}
\name{dlPFC.TOD}
\alias{dlPFC.TOD}
\title{Dorsolateral Prefrontal Cortex Data (control subjects) Time of Death}
\format{
A time of death vector.
}
\usage{
dlPFC.TOD
}
\description{
Time of death vectors from human postmortem brain data (dorsolateral prefrontal cortex).
}
\references{
Seney, M. L., Cahill, K., Enwright, J. F., Logan, R. W., Huo, Z., Zong, W., ... & McClung, C. A. (2019). Diurnal rhythms in gene expression in the prefrontal cortex in schizophrenia. Nature communications, 10(1), 1-11.
}
\keyword{datasets}
