% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_MC_powerCircadian_t.R
\name{F_MC_powerCircadian_t}
\alias{F_MC_powerCircadian_t}
\title{F-statistics-based Monte-Carlo Circadian Power Calculation assuming independent t-distributed errors}
\usage{
F_MC_powerCircadian_t(
  B_MC = 10000,
  n,
  A,
  df,
  phi = 0,
  period = 24,
  C = 10,
  cts = NULL,
  alpha = 0.05
)
}
\arguments{
\item{B_MC}{Numer of Monte-Carlo simulations. Default is 10000.}

\item{n}{Sample size.}

\item{A}{Amplitude of the sine curve: \eqn{A * sin(2\pi/period * (phi + cts)) + C}.}

\item{df}{Degree of freedom in the independent t-distributed error term \eqn{t(df)}.}

\item{phi}{Phase of the sine curve. Default is 0.}

\item{period}{Period of the sine curve. Default is 24.}

\item{C}{Offset of the sine curve. Default is 10.}

\item{cts}{Circadian time design vector.}

\item{alpha}{Type I error control. Default is 0.05.}
}
\value{
A vector of empirical type I error and Monte-Carlo power based on F-statistics.
}
\description{
Monte-Carlo circadian power calculation based on F-statistics assuming independent t-distributed errors
}
\details{
Calculate power of circaidan data from Monte-Carlo simulated data assuming independent t-distributed errors.
}
\examples{
B_MC = 10000
n = 100
cts = seq(0,24,length.out = n+1)[-1]
A = 1
df = 3
phi = 0
C = 10
F_MC_powerCircadian_t(B_MC, n, A, df, phi=0, period = 24, cts=cts, alpha = 0.05)
}
\author{
Wei Zong, Zhiguang Huo
}
